pyproject.toml
[project]
name = "rps-plus-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "python-dotenv>=1.2.1",
]
